requirements.txt
async-tkinter-loop==0.9.3
asyncgui==0.6.3
asyncgui-ext-clock==0.3.1
asynctkinter==0.4.1
beautifulsoup4==4.12.3
certifi==2024.12.14
charset-normalizer==3.4.1
customtkinter==5.2.2
darkdetect==0.8.0
idna==3.10
MouseInfo==0.1.3
packaging==24.2
pillow==11.0.0
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python3-xlib==0.15
pytweening==1.2.0
requests==2.32.3
soupsieve==2.6
typing_extensions==4.12.2
urllib3==2.3.0
yt-dlp==2024.12.23
